Rename types module to wrappers
