+++ FLIR/conservator/managers/__init__.py
@@ -3,7 +3,7 @@ A Manager is simply a bundle of utilities for querying a specific type.
 """
 from FLIR.conservator.managers.media import MediaTypeManager
 from FLIR.conservator.managers.searchable import SearchableTypeManager
-from FLIR.conservator.types import Collection, Dataset, Project, Video, Image
+from FLIR.conservator.wrappers import Collection, Dataset, Project, Video, Image
 
 
 class CollectionManager(SearchableTypeManager):
@@ -49,7 +49,7 @@ class DatasetManager(SearchableTypeManager):
         super().__init__(conservator, Dataset)
 
     def from_local_path(self, path="."):
-        """Create a :class:`~FLIR.conservator.types.Dataset` from a `path`
+        """Create a :class:`~FLIR.conservator.wrappers.Dataset` from a `path`
         containing an ``index.json`` file."""
         return self._underlying_type.from_local_path(self._conservator, path)
