+++ FLIR/conservator/wrappers/dataset.py
@@ -11,9 +11,9 @@ from FLIR.conservator.generated.schema import (
     DeleteDatasetInput,
 )
 from FLIR.conservator.paginated_query import PaginatedQuery
-from FLIR.conservator.types.dataset_frame import DatasetFrame
-from FLIR.conservator.types.type_proxy import requires_fields
-from FLIR.conservator.types.queryable import QueryableType
+from FLIR.conservator.wrappers.dataset_frame import DatasetFrame
+from FLIR.conservator.wrappers.type_proxy import requires_fields
+from FLIR.conservator.wrappers.queryable import QueryableType
 from FLIR.conservator.util import download_files
