+++ FLIR/conservator/wrappers/image.py
@@ -1,6 +1,6 @@
 from FLIR.conservator.generated import schema
-from FLIR.conservator.types.type_proxy import requires_fields
-from FLIR.conservator.types.queryable import QueryableType
+from FLIR.conservator.wrappers.type_proxy import requires_fields
+from FLIR.conservator.wrappers.queryable import QueryableType
 from FLIR.conservator.util import download_file
